- Fixed compiler warnings - Mage improvement: Mage now using Berserking after flee - Mage improvement: (WOTLK) Mage now using Molten Armor instead of Mage Armor

+++ playerbot/strategy/triggers/GlyphTriggers.h
@@ -22,7 +22,7 @@ namespace ai
             if (requiredSpellId != -1 && !bot->HasSpell(requiredSpellId))
                 return false;
 
-            if (requiredLevel != -1 && bot->GetLevel() < requiredLevel)
+            if (requiredLevel != -1 && (int)bot->GetLevel() < requiredLevel)
                 return false;
 
             if (!requiredCombatStrategy.empty() && !ai->HasStrategy(requiredCombatStrategy, BotState::BOT_STATE_COMBAT))
@@ -53,7 +53,7 @@ namespace ai
             if (requiredSpellId != -1 && !bot->HasSpell(requiredSpellId))
                 return true;
 
-            if (requiredLevel != -1 && bot->GetLevel() < requiredLevel)
+            if (requiredLevel != -1 && (int)bot->GetLevel() < requiredLevel)
                 return true;
 
             if (!requiredCombatStrategy.empty() && !ai->HasStrategy(requiredCombatStrategy, BotState::BOT_STATE_COMBAT))
